+++ CHANGES.md
@@ -4,6 +4,8 @@
 ----------------------------
 
 * fix odt image inclusion
+* Add anchor image_url_load, image_load
+* Replace width and height for image_url_load, image_load by max_width and max_height
 
 
 1.2.17          (2019-11-07)